rename workspaces to workspace
