+++ packages/core/src/index.ts
@@ -31,7 +31,7 @@ export * from './features/cohort';
 export * from './features/filters';
 export * from './features/guppy';
 export * from './features/aiSearch';
-export * from './features/workspaces';
+export * from './features/workspace';
 export * from './features/submission';
 
 export {
