+++ packages/frontend/src/pages/Workspace/data.ts
@@ -2,7 +2,7 @@ import { GetServerSideProps } from 'next';
 import { getNavPageLayoutPropsFromConfig } from '../../lib/common/staticProps';
 import ContentSource from '../../lib/content';
 import { GEN3_COMMONS_NAME } from '@gen3/core';
-import { WorkspaceConfiguration } from '../../features/Workspaces';
+import { WorkspaceConfiguration } from '../../features/Workspace';
 import { WorkspacePageProps } from './types';
 
 export const WorkspacesPageGetServerSideProps: GetServerSideProps<
