+++ packages/frontend/src/pages/Workspace/types.ts
@@ -1,5 +1,5 @@
 import { NavPageLayoutProps } from '../../features/Navigation';
-import { WorkspaceConfiguration } from '../../features/Workspaces';
+import { WorkspaceConfiguration } from '../../features/Workspace';
 
 
 export interface WorkspacePageProps extends NavPageLayoutProps {
